# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindTrace",
    "description": "Service for managing mindtrace",
    "version": "3.1.3"
  },
  "paths": {
...
    "/v1/auth/refresh/": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Невалидный или истёкший refresh-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.invalid_refresh_token",
                  "message": "Невалидный или истёкший refresh-токен",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "internal",
                  "message": "Внутренняя ошибка сервера",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/email/send-verification/": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Email Verification",
        "operationId": "send_email_verification",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Невалидный или истёкший access-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.invalid_access_token",
                  "message": "Невалидный или истёкший access-токен",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Учётная запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.user_credentials_not_found",
                  "message": "Учётная запись не найдена",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "409": {
            "description": "Email уже подтверждён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.email_already_verified",
                  "message": "Email уже подтверждён",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "429": {
            "description": "Cooldown повторной отправки кода ещё не истёк",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.challenge_resend_cooldown",
                  "message": "Слишком частая отправка кода, повторите попытку позже",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "internal",
                  "message": "Внутренняя ошибка сервера",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/email/verify/": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Email",
        "operationId": "verify_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Неверный код подтверждения",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.verification_code_invalid",
                  "message": "Неверный код подтверждения",
                  "details": {
                    "field": "code"
                  },
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "Невалидный или истёкший access-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.invalid_access_token",
                  "message": "Невалидный или истёкший access-токен",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Учётная запись или активный challenge не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "auth.user_credentials_not_found": {
                    "summary": "Учётная запись не найдена",
                    "value": {
                      "code": "auth.user_credentials_not_found",
                      "message": "Учётная запись не найдена",
                      "timestamp": "2024-01-01T12:00:00Z"
                    }
                  },
                  "auth.challenge_not_found": {
                    "summary": "Активный challenge отсутствует",
                    "value": {
                      "code": "auth.challenge_not_found",
                      "message": "Код подтверждения не запрошен или больше недоступен",
                      "timestamp": "2024-01-01T12:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email уже подтверждён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.email_already_verified",
                  "message": "Email уже подтверждён",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "410": {
            "description": "Срок действия кода истёк",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.challenge_expired",
                  "message": "Срок действия кода подтверждения истёк",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "429": {
            "description": "Превышено количество попыток ввода кода",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "auth.challenge_attempts_exceeded",
                  "message": "Превышено количество попыток ввода кода",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "internal",
                  "message": "Внутренняя ошибка сервера",
                  "timestamp": "2024-01-01T12:00:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код ошибки для программной обработки",
            "examples": [
              "invalid_input",
              "not_found",
              "internal"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Человекочитаемое сообщение об ошибке",
            "examples": [
              "Некорректный запрос",
              "Ресурс не найден"
            ]
          },
          "details": {
            "$ref": "#/components/schemas/OptionalDict",
            "description": "Дополнительные детали ошибки (опционально)",
            "examples": [
              {
                "field": "email",
                "reason": "Invalid email format"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Временная метка возникновения ошибки (UTC)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Базовая схема ответа ошибки API.\n\nИспользуется для всех типов ошибок и автоматически генерирует документацию OpenAPI.",
        "example": {
          "code": "invalid_input",
          "message": "Некорректный запрос",
          "timestamp": "2024-01-01T12:00:00Z"
        }
      },
...
      "OptionalDict": {
        "anyOf": [
          {
            "additionalProperties": true,
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "accessToken"
        ],
        "title": "TokenResponse"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "VerifyEmailRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}